# [WIP] sparse: use staged cursor and upper_bound for WAND