Use staged cursor and upper_bound to avoid recalculating the score from cursor 0 every time in the loop. And modify some variable names to be more readable.

Simple benchmark result(time used to search all queries once on the full dataset) with drop ratio build 0.32 and drop ratio search 0.6:

| BM25 |Before|After|Improvement %|
|---|---|---|---|
|[bigann SPLADE vectors of MSMarco](https://github.com/harsha-simhadri/big-ann-benchmarks/blob/main/dataset_preparation/sparse_dataset.md)|81,980ms|66,205ms|19.2%|
|Milvus 2.5 Full Text Search Vectors of MSMarco|1,005ms|1,034ms|~|

| IP |Before|After|Improvement %|
|---|---|---|---|
|[bigann SPLADE vectors of MSMarco](https://github.com/harsha-simhadri/big-ann-benchmarks/blob/main/dataset_preparation/sparse_dataset.md)|85,527ms|75,366ms|12%|
|Milvus 2.5 Full Text Search Vectors of MSMarco(using IP/TF-IDF metric though, not BM25)|2,839ms|3,191ms|-12%|

This optimization seems to have an improvement on learned sparse embedding like SPLADE, which is denser, but leads to a little bit of performance degration on more sparse embeddings.

Hold the review of this PR for now until more investigation.