@sparknack 2024-12-03
issue: #967