@sparknack 2025-02-19
> @sparknack is it good to review the PR?

since maxscore algorithum is better than wand, this optimization is downgraded.